Refactor components.
