+++ src/screens/Library/LibraryScreenHeader/index.js
@@ -1,0 +1,1 @@
+export * from './component';
